ultima clase

+++ ProyectoCiclo3.App.Dominio/Entidades/Buses.cs
@@ -5,7 +5,7 @@ namespace ProyectoCiclo3.App.Dominio{
         public int id {get;set;} 
         public string marca {get;set;}            
         public string modelo {get;set;}            
-        public float kilometraje {get;set;}      
+        public int kilometraje {get;set;}      
         public int numer_Asientos {get;set;}      
         public string placa {get;set;}
